requirements.txt
# Core LangChain and LangGraph
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.59
langchain-core==0.3.28

# FastAPI and async support
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sse-starlette==2.2.1

# Database
pg8000
sqlalchemy==2.0.36
alembic==1.14.0
supabase==2.11.0
pgvector==0.3.6

# OpenAI
openai==1.59.5
tiktoken==0.8.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Document processing
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.14

# Logging and monitoring
loguru==0.7.3
